# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Physical Venue Risk Assessment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List all assessments
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentSummary"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create a new assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentSummary"
...
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get a single assessment with detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetail"
        "404":
          description: Not found
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteAssessment
      tags: [assessments]
      summary: Delete an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /assessments/{id}/approve:
    post:
      operationId: approveAssessment
      tags: [assessments]
      summary: Approve an assessment (creates version snapshot)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveAssessmentInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentDetail"
        "404":
          description: Not found
  /assessments/{id}/versions:
    get:
      operationId: listAssessmentVersions
      tags: [assessments]
      summary: List version history for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentVersion"
  /assessments/{id}/audit-log:
    get:
      operationId: listAuditLog
      tags: [assessments]
      summary: List audit log entries for an assessment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
...
components:
  schemas:
...
    AssessmentSummary:
      type: object
      required: [id, title, status, version, createdAt, updatedAt]
      properties:
        id:
          type: integer
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        venueCity:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, under_review, approved, monitoring, review_required, escalated, archived]
        version:
          type: integer
        overallRating:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AssessmentDetail:
      type: object
      required: [id, title, status, version, risks, createdAt, updatedAt]
      properties:
        id:
          type: integer
        venueId:
          type: ["integer", "null"]
        venueName:
          type: ["string", "null"]
        venueCity:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, under_review, approved, monitoring, review_required, escalated, archived]
        version:
          type: integer
        intelSummary:
          type: ["string", "null"]
        analystNotes:
          type: ["string", "null"]
        overallRating:
          type: ["string", "null"]
        riskMatrix:
          $ref: "#/components/schemas/RiskMatrix"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/Risk"
        approvedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AssessmentInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        venueId:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [draft, under_review, approved, monitoring, review_required, escalated, archived]
        intelSummary:
          type: string
        analystNotes:
          type: string
    AssessmentUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        venueId:
          type: integer
        description:
          type: string
        status:
          type: string
          enum: [draft, under_review, approved, monitoring, review_required, escalated, archived]
        intelSummary:
          type: string
        analystNotes:
          type: string
    ApproveAssessmentInput:
      type: object
      required: [userId, changeSummary]
      properties:
        userId:
          type: integer
        changeSummary:
          type: string
    AssessmentVersion:
      type: object
      required: [id, assessmentId, version, changeSummary, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        version:
          type: integer
        changeSummary:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AuditLogEntry:
      type: object
      required: [id, action, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: ["integer", "null"]
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        action:
          type: string
        fieldChanged:
          type: ["string", "null"]
        oldValue:
          type: ["string", "null"]
        newValue:
          type: ["string", "null"]
        reason:
          type: ["string", "null"]
        createdAt:
          type: string
    RiskMatrix:
      type: object
      required: [id, assessmentId, areaRisk, accessControl, arrivalDeparture, parking, personnel, medical, hse, extraction, overallRating]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        areaRisk:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        accessControl:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        arrivalDeparture:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        parking:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        personnel:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        medical:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        hse:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        extraction:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        overallRating:
          type: string
          enum: [low, moderate, moderate_high, high, unknown]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    Risk:
      type: object
      required: [id, assessmentId, title, likelihood, impact, riskScore, category, createdAt, updatedAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [operational, financial, strategic, compliance, reputational, technical, other]
        likelihood:
          type: integer
          minimum: 1
          maximum: 5
        impact:
          type: integer
          minimum: 1
          maximum: 5
        riskScore:
          type: integer
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        mitigation:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        status:
          type: string
          enum: [open, mitigated, accepted, closed]
        createdAt:
          type: string
        updatedAt:
          type: string
...