# lib/api-zod/src/generated/types/assessmentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Physical Venue Risk Assessment API
 * OpenAPI spec version: 0.2.0
 */
import type { AssessmentInputStatus } from './assessmentInputStatus';

export interface AssessmentInput {
  /** @minLength 1 */
  title: string;
  venueId?: number;
  description?: string;
  status?: AssessmentInputStatus;
  intelSummary?: string;
  analystNotes?: string;
}